imprv: Gtk.Template stubs for GTK+3
This backports the type stubs from GTK4 for `Gtk.Template` to GTK3.
Most notably this now correctly marks `Gtk.Template` and
`Gtk.Template.Callback` so that they can be used as decorators.

Note: I have checked the documentation and ran some tests and
am pretty certain the interface is the same as in GTK 4, but somebody
else should probably double check.