typing: GObject: Type Properties using inference
This seems to work consistently, automatically
unless you rely on the implicit nullability of a property.
In that case, you will need to manually type it as:

```
prop = GObject.Property[T | None](type=T)
```

However, given that this behavior is ignored in other places
and that the alternative would be wildly more annoying
to deal with, it should probably be fine.